Business Requirements Document (BRD)
Vinaly is a specialized provider of printing services/products for individuals and businesses nationwide. With the advancement of technology and the trend of online ordering, the company aims to enhance the Vinaly.vn website to ensure the best experience for customers when making online purchases.

This document is intended to outline the requirements for improving the functionality of the company's e-commerce website, targeting users who have a demand for ordering and designing custom printed gift products. The document is drafted by the PQL project team for the Vinaly company's management board and relevant parties to review and approve the project.